backend/requirements.txt
# BitCred Backend Dependencies

# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.0

# HTTP client
httpx==0.27.0

# Starknet
starknet-py==0.23.0

# AI / Scoring
numpy==1.26.4
scikit-learn==1.5.0

# Utilities
python-dotenv==1.0.1